Refactor RaftConsensus to inherit from abstract base class
