+++ projects/libs/core-library-typescript/src/consensus/consensus.ts
@@ -3,27 +3,28 @@ import { types } from "../common-library/types";
 
 import { Gossip } from "../gossip";
 
-import { RaftConsensusConfig, RaftConsensus } from "./raft-consensus";
+import { RaftConsensusConfig, BaseConsensus } from "./base-consensus";
+import { RaftConsensus } from "./raft-consensus";
 
 export class Consensus {
-  private raftConsensus: RaftConsensus;
+  private actualConsensus: BaseConsensus;
 
   constructor(
     config: RaftConsensusConfig, gossip: types.GossipClient,
     virtualMachine: types.VirtualMachineClient, blockStorage: types.BlockStorageClient,
      transactionPool: types.TransactionPoolClient) {
-    this.raftConsensus = new RaftConsensus(config, gossip, blockStorage, transactionPool, virtualMachine);
+    this.actualConsensus = new RaftConsensus(config, gossip, blockStorage, transactionPool, virtualMachine);
   }
 
   async initialize() {
-    return this.raftConsensus.initialize();
+    return this.actualConsensus.initialize();
   }
 
   async shutdown() {
-    return this.raftConsensus.shutdown();
+    return this.actualConsensus.shutdown();
   }
 
   async gossipMessageReceived(fromAddress: string, messageType: string, message: any) {
-    await this.raftConsensus.onMessageReceived(fromAddress, messageType, message);
+    await this.actualConsensus.onMessageReceived(fromAddress, messageType, message);
   }
 }
